ICU-23040 Extract inner classes as standalone
